+++ workspaces/ballerina-extension/src/core/extension.ts
@@ -21,13 +21,13 @@ import { join, sep } from 'path';
 import { exec, spawnSync } from 'child_process';
 import { LanguageClientOptions, State as LS_STATE, RevealOutputChannelOn, ServerOptions } from "vscode-languageclient/node";
 import { getServerOptions } from '../server/server';
-import { ExtendedLangClient } from './extended-language-client';
+import { BallerinaProject, ExtendedLangClient } from './extended-language-client';
 import { debug, log, getOutputChannel, outputChannel, isWindows, isSupportedVersion, VERSION } from '../utils';
 import { AssertionError } from "assert";
 import {
     BALLERINA_HOME, ENABLE_ALL_CODELENS, ENABLE_TELEMETRY, ENABLE_SEMANTIC_HIGHLIGHTING, OVERRIDE_BALLERINA_HOME,
     BALLERINA_LOW_CODE_MODE, ENABLE_PERFORMANCE_FORECAST, ENABLE_DEBUG_LOG, ENABLE_BALLERINA_LS_DEBUG,
-    ENABLE_CONFIGURABLE_EDITOR, ENABLE_EXPERIMENTAL_FEATURES, ENABLE_NOTEBOOK_DEBUG, ENABLE_RUN_FAST
+    ENABLE_EXPERIMENTAL_FEATURES, ENABLE_NOTEBOOK_DEBUG, ENABLE_RUN_FAST
 }
     from "./preferences";
 import TelemetryReporter from "vscode-extension-telemetry";
@@ -562,11 +562,6 @@ export class BallerinaExtension {
         return <boolean>workspace.getConfiguration().get(ENABLE_PERFORMANCE_FORECAST);
     }
 
-    public isConfigurableEditorEnabled(): boolean {
-        return this.isCodeServerEnv() ||
-            <boolean>workspace.getConfiguration().get(ENABLE_CONFIGURABLE_EDITOR);
-    }
-
     public enabledExperimentalFeatures(): boolean {
         return <boolean>workspace.getConfiguration().get(ENABLE_EXPERIMENTAL_FEATURES);
     }
@@ -649,6 +644,7 @@ class DocumentContext {
     private editorChangesCallbacks: Array<(change: Change) => void> = [];
     private latestDocument: Uri | undefined;
     private activeDiagram: boolean = false;
+    private ballerinProject: BallerinaProject;
 
     public diagramTreeElementClicked(construct: ConstructIdentifier): void {
         this.diagramTreeElementClickedCallbacks.forEach((callback) => {
@@ -675,6 +671,15 @@ class DocumentContext {
             return;
         }
         this.latestDocument = uri;
+    }
+
+    public setCurrentProject(ballerinProject: BallerinaProject) {
+        commands.executeCommand('setContext', 'isBallerinaProject', true);
+        this.ballerinProject = ballerinProject;
+    }
+
+    public getCurrentProject(): BallerinaProject {
+        return this.ballerinProject;
     }
 
     public getLatestDocument(): Uri | undefined {