+++ workspaces/ballerina-extension/src/project/activator.ts
@@ -18,6 +18,7 @@ import { activatePasteJsonAsRecord } from "./cmds/json-to-record";
 import { activatePackCommand } from "./cmds/pack";
 import { activateRenameCommand } from "./cmds/rename";
 import { activateExtractCommand } from "./cmds/extract";
+import { activateConfigRunCommand } from "./cmds/configRun";
 
 export * from "./cmds/cmd-runner";
 
@@ -33,6 +34,9 @@ export function activate() {
 
     // activate ballerina run command
     activateRunCommand();
+
+    // activate ballerina run command in config
+    activateConfigRunCommand();
 
     // activate ballerina run command
     activateRunCmdCommand();